+++ tests/test_aggregation.py
@@ -10,6 +10,7 @@ import sys
 
 import numpy as np
 import pandas as pd
+import pandas.testing as pdt
 import pytest
 
 TESTPATH = os.path.dirname(os.path.abspath(__file__))
@@ -63,6 +64,7 @@ def test_aggreation_regions():
     )
 
     assert ["a", "b"] == io_agg_with_names.get_regions().tolist()
+    assert io_agg_with_names.unit.index.equals(io_agg_with_names.Z.index)
 
     io_vec1 = io.aggregate(region_agg=reg_agg_vec1, inplace=False)
     io_vec2 = io.aggregate(region_agg=reg_agg_vec2, inplace=False)
@@ -75,6 +77,94 @@ def test_aggreation_regions():
     np.testing.assert_allclose(manual_agg.values, io_df.emissions.D_cba_reg.a)
 
     assert ["a", "b"] == io_df.get_regions().tolist()
+
+    # Testing the aggregation of duplicate regions/sectors
+
+    reg_rename_dict = reg_agg_df.set_index("original").squeeze().to_dict()
+    io.rename_regions(reg_rename_dict)
+    io_agg = io.aggregate_duplicates(inplace=False)
+    io.aggregate_duplicates(inplace=True)
+
+    pdt.assert_frame_equal(io.Z, io_agg.Z)
+    pdt.assert_frame_equal(io.Z, io_agg_with_names.Z)
+    pdt.assert_frame_equal(io.Y, io_agg.Y)
+    pdt.assert_frame_equal(io.Y, io_agg_with_names.Y)
+    pdt.assert_frame_equal(io.x, io_agg.x)
+    pdt.assert_frame_equal(io.x, io_agg_with_names.x)
+
+    pdt.assert_frame_equal(io.emissions.F, io_agg.emissions.F)
+    pdt.assert_frame_equal(io.emissions.F, io_agg_with_names.emissions.F)
+    pdt.assert_frame_equal(io.emissions.D_cba, io_agg.emissions.D_cba)
+    pdt.assert_frame_equal(io.emissions.D_cba, io_agg_with_names.emissions.D_cba)
+    pdt.assert_frame_equal(io.emissions.F_Y, io_agg.emissions.F_Y)
+    pdt.assert_frame_equal(io.emissions.F_Y, io_agg_with_names.emissions.F_Y)
+
+    pdt.assert_frame_equal(io.factor_inputs.F, io_agg.factor_inputs.F)
+    pdt.assert_frame_equal(io.factor_inputs.F, io_agg_with_names.factor_inputs.F)
+    pdt.assert_frame_equal(io.factor_inputs.D_cba, io_agg.factor_inputs.D_cba)
+    pdt.assert_frame_equal(
+        io.factor_inputs.D_cba, io_agg_with_names.factor_inputs.D_cba
+    )
+
+    assert io.unit.index.equals(io.Z.index)
+    assert io_agg.unit.index.equals(io_agg.Z.index)
+
+
+def test_aggreation_sectors():
+    """Test different possibilities to aggregate sectors"""
+
+    sec_agg_df = pd.DataFrame(
+        data=[
+            ("food", "eat"),
+            ("mining", "dig"),
+            ("manufactoring", "build"),
+            ("electricity", "util"),
+            ("construction", "build"),
+            ("trade", "marg"),
+            ("transport", "marg"),
+            ("other", "misc"),
+        ],
+        columns=["original", "aggregated"],
+    )
+    io = pymrio.load_test()
+    io_agg = io.aggregate(sector_agg=sec_agg_df, inplace=False)
+
+    test_sectors = ["manufactoring", "construction"]
+    assert (
+        io.Z.loc["reg2", "reg2"].loc[test_sectors, test_sectors].sum().sum()
+        == io_agg.Z.loc["reg2", "reg2"].loc["build", "build"]
+    )
+    assert (
+        io.Y.loc["reg2", "reg2"].loc[test_sectors, :].sum().sum()
+        == io_agg.Y.loc["reg2", "reg2"].loc["build", :].sum().sum()
+    )
+
+    test_rename = [("reg3", "other"), ("reg3", "misc")]
+    assert (
+        io.Z.loc[test_rename[0], test_rename[0]]
+        == io_agg.Z.loc[test_rename[1], test_rename[1]]
+    )
+
+    sec_rename_dict = sec_agg_df.set_index("original").squeeze().to_dict()
+    io.rename_regions(sec_rename_dict)
+    io.aggregate_duplicates(inplace=True)
+
+    io.calc_all()
+    io_agg.calc_all()
+
+    pdt.assert_frame_equal(io.Z, io_agg.Z)
+    pdt.assert_frame_equal(io.Y, io_agg.Y)
+    pdt.assert_frame_equal(io.x, io_agg.x)
+
+    pdt.assert_frame_equal(io.emissions.F, io_agg.emissions.F)
+    pdt.assert_frame_equal(io.emissions.D_cba, io_agg.emissions.D_cba)
+    pdt.assert_frame_equal(io.emissions.F_Y, io_agg.emissions.F_Y)
+
+    pdt.assert_frame_equal(io.factor_inputs.F, io_agg.factor_inputs.F)
+    pdt.assert_frame_equal(io.factor_inputs.D_cba, io_agg.factor_inputs.D_cba)
+
+    assert io.unit.index.equals(io.Z.index)
+    assert io_agg.unit.index.equals(io_agg.Z.index)
 
 
 def test_numerical_aggreation_sectors():